move theme to theme.js
